issue #1983 - move verification methods into unit test
Moved PopulateParameterNames.verify() to PopulateParameterNamesTest and
set the FHIRRegistry dependency to `test` scope.

Moved PopulateResourceTypes.verify() to PopulateResourceTypesTest just
for consistency.
